+++ CHANGELOG.md
@@ -3,6 +3,7 @@ v0.3.0 (in development)
 - Support Python 3.14
 - Drop support for Python 3.8 and 3.9
 - Change `EntryPoint` from a `NamedTuple` to a dataclass
+- Remove `EntryPoint.object` (deprecated in v0.2.0)
 
 v0.2.1 (2024-12-01)
 -------------------